pyproject.toml
[tool.poetry]
name = "palsta23"
version = "0.1.0"
description = "Tietokantasovellus harjoitustyö"
authors = ["HeljaeRaeisaenen <94612974+HeljaeRaeisaenen@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
flask = "^2.2.3"
flask-sqlalchemy = "^3.0.3"
psycopg2 = "^2.9.5"
python-dotenv = "^0.21.1"
invoke = "^2.0.0"


[tool.poetry.group.dev.dependencies]
pylint = "^2.17.0"
autopep8 = "^2.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
